fix: update line count method signatures and API consistency
- Change line_count() return type from u64 to usize for consistency
- Update TokenSource trait to use immutable reference for line_count()
- Fix lexer method name: get_current_line() -> get_current_line_count()
- Remove unused import in tests module
- Ensure API consistency across all components

+++ frontend/src/parser/core.rs
@@ -44,7 +44,7 @@ impl<'a> Parser<'a> {
         self.token_provider.advance();
     }
 
-    pub fn line_count(&mut self) -> u64 {
+    pub fn line_count(&mut self) -> usize {
         self.token_provider.line_count()
     }
